separating date picker in multiple components
